Merge join-order-1674-2
Author: exarkun
Reviewer: glyph
Fixes #1674

Order tables in the FROM clause of generated SQL statements in Axiom by
the order of Item subclasses which appear in the comparisons used to
generate those queries.  Also add a new IComparison implementation which
allows application-level code to explicitly specify the order of types
in the join.


git-svn-id: http://divmod.org/svn/Divmod/trunk@9850 866e43f7-fbfc-0310-8f2a-ec88d1da2979