Cross-transaction query iterators.
This adds a new method, {{{ItemQuery.paginate}}}, which splits a query's
result-gathering work into multiple "pages" so that we can deal with extremely
large result sets.  In the POP3 case, we will be going through every Message in
a user's store, in storeID order. This pagination is now done asynchronously so
that it does not block the entire server while a user's mailbox is loaded. (As
always, in JP's case, that was quite serious.)

Author: glyph

Reviewer: exarkun

Fixes #1823


git-svn-id: http://divmod.org/svn/Divmod/trunk@10861 866e43f7-fbfc-0310-8f2a-ec88d1da2979