Merge jsunit-1729
 * Author: jml
 * Reviewer: exarkun
 * Fixes #1729

Unit testing framework for JavaScript tests that can be run from Trial.


This branch provides a Divmod.Class-based JavaScript unit testing framework
built using TDD along the pattern provided by xUnit.

It is hoped that this code can be extended to provide better support for
testing LiveWidgets, and that the Trial support code can be improved in
future iterations.


git-svn-id: http://divmod.org/svn/Divmod/trunk@11210 866e43f7-fbfc-0310-8f2a-ec88d1da2979