Fix {{{asAccessibleTo}}} to yield the correct results in the face of a limit.
Fixes #2189

Author: glyph, washort

Reviewer: moe

This change also adds the very beginnings of an explicit, supported query-introspection API that works with all types of queries.  Currently it only works with limit() (the only thing that was necessary for this change) but future query-manipulation logic should be added to the IQuery interface so that code will work when presented with, e.g. an AttributeQuery or one of the 'distinct' query wrappers.


git-svn-id: http://divmod.org/svn/Divmod/trunk@13104 866e43f7-fbfc-0310-8f2a-ec88d1da2979