Remove ClickChronicle
ClickChronicle development isn't going anywhere and isn't likely to in
the foreseeable future.  The cost of ongoing maintenance overcomes the value
it is providing, so it is being abandoned.

There is no corresponding ticket for this change.  It is a kind of event
which is happens rarely enough that the development workflow does not
account for it.


git-svn-id: http://divmod.org/svn/Divmod/trunk@13258 866e43f7-fbfc-0310-8f2a-ec88d1da2979