Re-organize and better document Imaginary codebase.
Author: glyph, tenth

Reviewer: washort

Fixes #2181

This change does a lot of stuff.  While it is the first step down a long road, it may now be possible to start showing this code to people who have not personally worked on it, with at least a dim hope that they might understand something.

This change removes redundant terminology: "commands" and "actions" are now unified both in the documentation and in the code into just "actions".

It updates the documentation: there is now a README.txt with actual english words in it instead of doc/RUNNING with incomprehensible shell log.  Several docstrings were added to important functionality, such as Thing creation.  The README also attempts to walk people through the infrastructure to the point where they can effectively create a server, rather than having to learn all about Mantissa first.

We also relocated code to more sensible places.  The rather central Thing creation functionality was moved into "imaginary.creation", out of the unfortunately named "quiche" module.  "npc" was renamed to "mice" since it included no general non-player-characater functionality.  Completely dead code, such as the 'places' module, has been deleted.

The decidedly half-baked example functionality, such as the actual application code for the quiche vending machine and hiragana mouse has been moved into the ExampleGame sub-project, to make the distinction between framework functionality (such as clothing) and examples or experiments in how to implement applications.

The mostly 'note to self' "TODO" has been replaced with a link to the ticket list.

There are also a few general cleanups, such as coding standard adjustments and fixes so that the codebase produces no warnings from 'pyflakes'.

That is not dead, which can eternal lie...


git-svn-id: http://divmod.org/svn/Divmod/trunk@13308 866e43f7-fbfc-0310-8f2a-ec88d1da2979