Merge athena-widget-plugin-2100: twistd athena-widget plugin displays traceback if the webserver is hit more than once
This was implemented by creating LivePage subclass
nevow._widget_plugin.WidgetPluginRoot; this is what is instantiated in
makeService() instead of ElementRenderingLivePage. At each request of this
resource, WidgetPluginRoot serves a new instance of ElementRenderingLivePage
which gets passed a new instance of the widget (element with which the pluggin
was called).

Author: glyph, oubiwann
Reviewer: exarkun
Fixes #2100


git-svn-id: http://divmod.org/svn/Divmod/trunk@13447 866e43f7-fbfc-0310-8f2a-ec88d1da2979