Merge repeated-edits-2307-2
Author: moe, exarkun
Reviewer: oubiwann
Fixes #2307

Change RepeatableFormParameter so that it coerces submissions to
RepeatableSubmission instances, with separate fields for creations,
modifications, and deletions, instead of a single list of all the
input.  Update the address book code which used repeatable form
parameters to account for this and add some acceptance tests which
exercise the functionality.


git-svn-id: http://divmod.org/svn/Divmod/trunk@13690 866e43f7-fbfc-0310-8f2a-ec88d1da2979