Allow customization of global client-side behavior via a client page class.
This change refactors most Athena global JavaScript state into a client-side PageWidget class, and allows that class to be specified by application code.  The page widget javascript class is responsible for, among other things, handling the "connection lost" notification, which allows for pages which intelligently reconnect or otherwise provide customized behavior in response to a connection-lost notification.  A brief interactive example of a page which can reconnect without alerting the user is included.

In addition to the refactorings, this change touches a lot of old code, and thereby adds a lot of unit test coverage to code which was either completely uncovered or covered only by nits.

Fixes #2310

Fixes #2334

Author: glyph, moe

Reviewer: exarkun


git-svn-id: http://divmod.org/svn/Divmod/trunk@14017 866e43f7-fbfc-0310-8f2a-ec88d1da2979