Cope with bogus email addresses in password reset.
Fixes #2357

Author: glyph

Reviewer: oubiwann

Two tracebacks were eliminated in the process of resetting your password based on an email address which did not exist in the login system.  However, this change does not provide any feedback to the user on a mistyped email address, to discourage attempts to harvest addresses or hack passwords.


git-svn-id: http://divmod.org/svn/Divmod/trunk@14058 866e43f7-fbfc-0310-8f2a-ec88d1da2979