Merge safer-mkbranch-2384-2
Author: exarkun
Reviewer: glyph, washort
Fixes #2384

Add a check to mkbranch so that it refuses to try to create a new branch with
the same name as an existing branch.  Previously this would put a directory
named trunk into the existing branch.  Now it will exit early without changing
the repository or any working copies.


git-svn-id: http://divmod.org/svn/Divmod/trunk@14239 866e43f7-fbfc-0310-8f2a-ec88d1da2979