Remove the long-lamented "start menu" button and many unnecessary tabs.
Fixes #2387

Author: glyph

Reviewer: moe

As an alternative in the one place (the "admin" application) where we still really need sub-tabs, this implements an ugly-but-functional dropdown menu that shows one level of tab descent (we no longer have '''any''' places that have sub-sub-tabs, so multiple levels are no longer implemented).

This change also removes the long defunct "ABC", "DEF" etc sub-tabs for the addressbook.


git-svn-id: http://divmod.org/svn/Divmod/trunk@14314 866e43f7-fbfc-0310-8f2a-ec88d1da2979