Merge cookie-domain-2487-2
Author: exarkun
Reviewer: glyph
Fixes #2487

Add a hook to nevow.guard.SessionWrapper which allows the domain
parameter of the session cookie to be specified.

Implement this hook in Mantissa's PersistentSessionWrapper to
base the value on the domains the site is configured with and a
flag indicating whether the cookie should be valid for subdomains.


git-svn-id: http://divmod.org/svn/Divmod/trunk@14865 866e43f7-fbfc-0310-8f2a-ec88d1da2979