Merge friendly-resolve-2601
Author: exarkun
Reviewer: glyph
Fixes: #2601

Change Thing.knownAs to consider names which match Thing.name case-
insensitively or which are word-boundary-delimited substrings of
Thing.name as matching, and so to return True for those cases.

This makes "get Foo" the same as "get foo" and will let things like
"get red" work for an object named "red shield".


git-svn-id: http://divmod.org/svn/Divmod/trunk@15651 866e43f7-fbfc-0310-8f2a-ec88d1da2979