Merge put-full-message-2602
Author: exarkun
Reviewer: glyph
Fixes: #2602

Change the message sent to an actor when they try to put something into a
container which will not fit in that container so that it explains why the
action failed.

Also remove an unclear test in favor of one which is more obvious as to its
intent (testPutSelf replaced with test_putFullFails).


git-svn-id: http://divmod.org/svn/Divmod/trunk@15658 866e43f7-fbfc-0310-8f2a-ec88d1da2979