Merge install-css-2882
Author: exarkun
Reviewer: jonathanj
Fixes: #2882

Add a test to Nevow which verifies that the `nevow.taglibrary.tabbedPane`
module can be imported.  At import time his module depends on a css file
which was not being installed by distutils.  Also, fix setup.py so that
this css file (and the rest of the `css/` hierarchy) is installed.


git-svn-id: http://divmod.org/svn/Divmod/trunk@17320 866e43f7-fbfc-0310-8f2a-ec88d1da2979