Merge axiomatic-reactor-selection-2864
Author: exarkun
Reviewer: glyph
Fixes: #2864
Fixes: #420

Fix `axiomatic start` so that if the `--reactor` option is passed, the given reactor
is installed before any application code is loaded.  This allows reactor selection
to work even if application code imports the reactor.

Also fix pidfile handling on Windows, and an `axiomatic start` test which was making
an invalid assertion on that platform.


git-svn-id: http://divmod.org/svn/Divmod/trunk@17530 866e43f7-fbfc-0310-8f2a-ec88d1da2979