Merge scheduler-nonitem-1408-2
Author: exarkun
Reviewer: glyph
Fixes: #1408

Introduce an in-memory-only `IScheduler` adapter for `Store` which supercedes
the `Scheduler` and `SubScheduler` item classes.  `Scheduler` and `SubScheduler`
still exist, but they will be deprecated shortly, and then upgraded out of
existence.  The correct way to get the scheduler for a store is to adapt the
store to `IScheduler`.  This will succeed without any particular database
state being created to support it.


git-svn-id: http://divmod.org/svn/Divmod/trunk@17723 866e43f7-fbfc-0310-8f2a-ec88d1da2979