Merge quotient-mxlookup-2941
Author: exarkun
Reviewer: glyph
Fixes: #2941

Fix the Quotient MX lookup code to work with Twisted 8.2 and newer by no
longer using the deprecated (and unfortunately, in a slight incompatibility
twist, disconnected) `protocol` attribute of the `Resolver`.  Fortunately,
this just involves deleting some code (but adding unit tests), as the behavior
the broken code was trying to provide is provided automatically by Twisted
now.


git-svn-id: http://divmod.org/svn/Divmod/trunk@17759 866e43f7-fbfc-0310-8f2a-ec88d1da2979